Added automatic neighbour list expansion and skin implementation
Previously the NeighbourList is a struct, now it is chnaged to a
mutable struct to allow on-the-fly increase of the maximum number
of supported neighbours.

A new keyword `skin` is added such that if used, the neighbour list
will only be rebuilt if any atom (wrapped in the primitive cell) has
moved more than this distance.